add subscriptions

+++ app/api/board/route.js
@@ -24,6 +24,14 @@ export async function POST(req) {
     await connectMongo();
 
     const user = await User.findById(session.user.id);
+
+    if (!user.hasAccess) {
+      return NextResponse.json(
+        { error: "Please subscribe first." },
+        { status: 403 }
+      );
+    }
+
     const board = await Board.create({
       name: body.name,
       userId: user._id,
@@ -56,9 +64,17 @@ export async function DELETE(req) {
       return NextResponse.json({ error: "Not Authorized." }, { status: 401 });
     }
 
+    const user = await User.findById(session?.user?.id);
+
+    if (!user.hasAccess) {
+      return NextResponse.json(
+        { error: "Please subscribe first." },
+        { status: 403 }
+      );
+    }
+
     await Board.deleteOne({ _id: boardId, userId: session?.user?.id });
 
-    const user = await User.findById(session?.user?.id);
     user.boards = user.boards.filter((id) => id.toString() !== boardId);
     await user.save();
